requirements.txt
# data science
numpy<2.0.0
pandas==2.2.3
matplotlib==3.10.6
seaborn==0.13.2
python-dotenv
python-multipart

# Uncomment if you use sklearn
scikit-learn==1.6.1

# Install the correct TensorFlow version
# Uncomment if you use TensorFlow
tensorflow==2.16.2    # For both Intel and Apple silicon
keras==3.8.0

# Google Cloud
# Uncomment if you use Google Cloud Storage
google-cloud-storage

# API
fastapi
uvicorn
pillow
